Ensuring only one visible slide at a time
"People with ludicrously tall monitors will just have to suffer," I
thought, trapped  for the summer in a rented bedroom without the normal
comforts of a home office. Then I tried to view my presentation on the
ludicrously tall monitor in my home office, and knew something had to be
done.